move test to interceptors/cache.js
